requirement.txt
python-telegram-bot==20.7
sentence-transformers==2.6.1
transformers==4.36.2
huggingface-hub==0.20.3
numpy==1.26.3
python-dotenv==1.0.0
torch>=2.0
